refactor: move user permission into file
